requirements.txt
ipykernel
numpy
pandas
scikit-learn
kagglehub
matplotlib
seaborn
python-dotenv
langchain-community
langchain-openai
langchain-huggingface
langchain-chroma
transformers
streamlit
notebook
ipywidgets
gradio
